+++ PlatonicRelationships/ModEntry.cs
@@ -1,3 +1,4 @@
+using System;
 using PlatonicRelationships.Framework;
 using StardewModdingAPI;
 using StardewModdingAPI.Events;
@@ -14,7 +15,7 @@ internal class ModEntry : Mod
     private ModConfig Config;
 
     /// <summary>Adjusts event data to disable vanilla 10-heart events for NPCs the player isn't dating.</summary>
-    private readonly EventEditor EventEditor = new EventEditor();
+    private EventEditor EventEditor;
 
 
     /*********
@@ -26,8 +27,14 @@ internal class ModEntry : Mod
         this.Config = helper.ReadConfig<ModConfig>();
         GamePatcher.Apply(this.ModManifest.UniqueID, this.Monitor);
 
+        DataModel data = this.LoadDataFile();
+
         if (this.Config.AddDatingRequirementToRomanticEvents)
+        {
+            this.EventEditor = new EventEditor(data, helper.GameContent.ParseAssetName);
+
             helper.Events.Content.AssetRequested += this.OnAssetRequested;
+        }
     }
 
 
@@ -37,7 +44,29 @@ internal class ModEntry : Mod
     /// <inheritdoc cref="IContentEvents.AssetRequested" />
     private void OnAssetRequested(object sender, AssetRequestedEventArgs e)
     {
-        if (this.EventEditor.CanEdit(e.NameWithoutLocale))
-            e.Edit(this.EventEditor.Edit);
+        this.EventEditor.OnAssetRequested(e);
+    }
+
+    /// <summary>Load the mod data from the <c>assets/data.json</c> file.</summary>
+    private DataModel LoadDataFile()
+    {
+        const string filePath = "assets/data.json";
+        try
+        {
+            DataModel data = this.Helper.Data.ReadJsonFile<DataModel>(filePath);
+            if (data is not null)
+            {
+                data.RomanticEvents ??= new();
+                return data;
+            }
+
+            this.Monitor.Log($"The '{filePath}' file is missing or empty, so Platonic Relationships may not work correctly. You can try reinstalling the mod to reset the file.");
+        }
+        catch (Exception ex)
+        {
+            this.Monitor.Log($"The '{filePath}' file couldn't be loaded, so Platonic Relationships may not work correctly. You can try reinstalling the mod to reset the file.\n\nTechnical details: {ex}");
+        }
+
+        return new DataModel();
     }
 }